add e2e-test for auto-tagging after export. refactoring: extract auto-tag e2e-tests to a new file
