+++ e2e/fixtures/fixtures.js
@@ -1,0 +1,6 @@
+export const isType = "module.exports = function isType() { return 'got is-type'; };";
+export const isString =
+  "const isType = require('./is-type.js'); module.exports = function isString() { return isType() +  ' and got is-string'; };";
+export const barFooFixture =
+  "const isString = require('../utils/is-string.js'); module.exports = function foo() { return isString() + ' and got foo'; };";
+export const appPrintBarFoo = "const barFoo = require('./components/bar/foo'); console.log(barFoo());";
